+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [2.3.0] - 2020-09-21
 
 ### Added
 
@@ -54,6 +54,7 @@
 
   * Initial release
 
+[2.3.0]: https://github.com/stjude/mtsg/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/stjude/mtsg/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/stjude/mtsg/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/stjude/mtsg/compare/v1.0.0...v2.0.0